# docs/api-documentations/openapi/asset.openapi.json
{
  "components": {
    "schemas": {
      "AgentRequestCreateDto": {
        "properties": {
          "idempotencyKey": {
            "description": "客户端为一次用户意图生成的幂等键；网络重试时保持不变",
            "maxLength": 128,
            "type": "string"
          },
          "items": {
            "description": "共享物品清单（同一库存条目整单只能一次）",
            "items": {
              "$ref": "#/components/schemas/AgentRequestItemDto"
            },
            "maxItems": 100,
            "type": "array"
          },
          "recipientIds": {
            "description": "受领人 id 列表（不能选择自己、不能重复；1～100 人）",
            "items": {
              "type": "number"
            },
            "type": "array"
          }
        },
        "required": [
          "recipientIds",
          "items"
        ],
        "type": "object"
      },
      "AgentRequestItemDto": {
        "properties": {
          "inventoryItemId": {
            "description": "库存条目 id",
            "minimum": 1,
            "type": "number"
          },
          "purpose": {
            "description": "用途（代交共享清单可空）",
            "maxLength": 200,
            "type": "string"
          },
          "qty": {
            "description": "清单总数量（正整数；不按受领人分摊）",
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "inventoryItemId",
          "qty"
        ],
        "type": "object"
      },
...
      "AssetCategoryBatchDeleteDto": {
        "properties": {
          "ids": {
            "description": "分类 id 列表（1～100 个，不重复）",
            "items": {
              "type": "number"
            },
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "type": "object"
      },
      "AssetCategoryCreateDto": {
        "properties": {
          "name": {
            "description": "分类名称",
            "maxLength": 50,
            "type": "string"
          },
          "parentId": {
            "description": "父分类 id（顶级分类为 null；业务只维护一级子分类）",
            "minimum": 1,
            "type": "number"
          },
          "sort": {
            "default": 0,
            "description": "排序（小的在前）",
            "type": "number"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "AssetCategoryUpdateDto": {
        "properties": {
          "name": {
            "description": "分类名称",
            "maxLength": 50,
            "type": "string"
          },
          "sort": {
            "default": 0,
            "description": "排序（小的在前）",
            "type": "number"
          },
          "status": {
            "description": "状态",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "status"
        ],
        "type": "object"
      },
...
      "AssetDictItemBatchDeleteDto": {
        "properties": {
          "ids": {
            "description": "字典项 id 列表（1～100 个，不重复）",
            "items": {
              "type": "number"
            },
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "type": "object"
      },
      "AssetDictItemCreateDto": {
        "properties": {
          "dictType": {
            "description": "字典类型",
            "enum": [
              "UNIT",
              "CHANGE_TYPE",
              "SUPPLIER",
              "BRAND",
              "SPEC",
              "ASSET_SPEC",
              "ASSET_MODEL"
            ],
            "type": "string"
          },
          "name": {
            "description": "字典项名称",
            "maxLength": 50,
            "type": "string"
          },
          "sort": {
            "default": 0,
            "description": "排序（小的在前）",
            "type": "number"
          }
        },
        "required": [
          "dictType",
          "name"
        ],
        "type": "object"
      },
      "AssetDictItemUpdateDto": {
        "properties": {
          "name": {
            "description": "字典项名称",
            "maxLength": 50,
            "type": "string"
          },
          "sort": {
            "default": 0,
            "description": "排序（小的在前）",
            "type": "number"
          },
          "status": {
            "description": "状态",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "status"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "csrfHeader": {
        "description": "CSRF 双提交头（状态变更请求必须与 wbme_csrf Cookie 一致）",
        "in": "header",
        "name": "X-WBME-CSRF-Token",
        "type": "apiKey"
      },
      "sessionCookie": {
        "description": "服务端会话 Cookie（HttpOnly；登录/激活成功后下发）",
        "in": "cookie",
        "name": "wbme_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "asset 部署单元（资产系统）OpenAPI 契约。构建期生成（主 PRD §9.5）：统一错误结构见 components.schemas.ErrorResponse；默认会话 Cookie + CSRF 头认证，公开路由无 security。",
    "title": "WBME asset API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/agent-requests": {
      "get": {
        "operationId": "AgentClaimController_listHistory",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "申请类型",
            "in": "query",
            "name": "requestType",
            "required": false,
            "schema": {
              "enum": [
                "CONSUMABLE_REQUEST",
                "AGENT_REQUEST"
              ],
              "type": "string"
            }
          },
          {
            "description": "审批状态",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "description": "发起人姓名（范围历史筛选）",
            "in": "query",
            "name": "applicantName",
            "required": false,
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentClaim"
        ]
      },
      "post": {
        "operationId": "AgentClaimController_submit",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRequestCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AgentClaim"
        ]
      }
    },
    "/api/v1/agent-requests/mine": {
      "get": {
        "operationId": "AgentClaimController_listMine",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "申请类型",
            "in": "query",
            "name": "requestType",
            "required": false,
            "schema": {
              "enum": [
                "CONSUMABLE_REQUEST",
                "AGENT_REQUEST"
              ],
              "type": "string"
            }
          },
          {
            "description": "审批状态",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "description": "发起人姓名（范围历史筛选）",
            "in": "query",
            "name": "applicantName",
            "required": false,
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentClaim"
        ]
      }
    },
...
    "/api/v1/categories": {
      "get": {
        "operationId": "CategoryController_list",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "状态过滤",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "ACTIVE",
                "DISABLED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      },
      "post": {
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCategoryCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "put": {
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCategoryUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/categories/batch": {
      "delete": {
        "operationId": "CategoryController_batchDelete",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCategoryBatchDeleteDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
...
    "/api/v1/dict-items": {
      "get": {
        "operationId": "DictController_list",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "字典类型过滤",
            "in": "query",
            "name": "dictType",
            "required": false,
            "schema": {
              "enum": [
                "UNIT",
                "CHANGE_TYPE",
                "SUPPLIER",
                "BRAND",
                "SPEC",
                "ASSET_SPEC",
                "ASSET_MODEL"
              ],
              "type": "string"
            }
          },
          {
            "description": "状态过滤",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "ACTIVE",
                "DISABLED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Dict"
        ]
      },
      "post": {
        "operationId": "DictController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetDictItemCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Dict"
        ]
      }
    },
    "/api/v1/dict-items/{id}": {
      "put": {
        "operationId": "DictController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetDictItemUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Dict"
        ]
      }
    },
    "/api/v1/dict-items/batch": {
      "delete": {
        "operationId": "DictController_batchDelete",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetDictItemBatchDeleteDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Dict"
        ]
      }
    },
...
  },
  "security": [
    {
      "csrfHeader": [],
      "sessionCookie": []
    }
  ],
  "servers": [],
...
}